requirements.txt
pytest==5.4.3
pytest-django==3.9.0
pytest-mock==3.2.0